Finished UI of sis3302, but checkboxes remain checked, even if the config is different. Cause yet unknown.

+++ sis3302module.h
@@ -48,8 +48,8 @@ public:
     unsigned int stop_delay;
     unsigned int nof_events;
 
-    uint8_t irq_level;
-    uint8_t irq_vector;
+    int irq_level;
+    int irq_vector;
 
     bool autostart_acq;
     bool internal_trg_as_stop;
